Do not import fixtures
