--- tests/components/generic_hygrostat/conftest.py
@@ -1,3 +1,0 @@
-"""Fixtures for generic_hygrostat tests."""
-
-from tests.components.switch.conftest import mock_switch_entities  # noqa: F401
