--- tests/components/generic_thermostat/conftest.py
@@ -1,3 +1,0 @@
-"""Fixtures for generic_thermostat tests."""
-
-from tests.components.switch.conftest import mock_switch_entities  # noqa: F401
